IR step 4c: the IR decides legality and modes
ir.Decide proves each counted integer operation inside its set (the signed
word, or U for the unsigned word's), which decides the refusal and, under
-checked, each operation's trap. gen and build refuse from it, the term
selection's checked renaming is no longer read, and the printers print the
decided function (FromFunc). Refusals name the source operation through
in-memory provenance (Stmt.Src), with binder names, in program order.

The domain became the term analysis's lattice: intervals over Z-bar with
128-bit ends in (-2^126, 2^126), so U and "bounded, not by this word" are
exact. New laws, each with its soundness argument (spec/ir.md 7.1): the
mask and the shift; unreachable arms and the bottom step; the step at one
continue; B = infinity; the unsigned order and u64 division as geometric.

The IR now also bounds alloc (table ...) elements and a map's cells, so
three differential cases drop "; checked:". Checked against executions: the
interpreter runs generated shapes and word- and shift-selected programs;
16 planted faults, all caught.

Emission: 8 changes (narrower Java scalars, uint32 limbs, windows bytes,
refusal text order), accepted. Compile time 0.93x: the decision runs once,
after the shift selection, and Finalize reuses its facts.

+++ gauntlet/check/testdata/emitted/examples_big_limbs.java.java
@@ -6,15 +6,15 @@ public final class examples_big_limbs.java {
 		int[] v2 = null;
 		int[] v7 = v3;
 		int v8 = 0;
-		long v9 = 1;
+		int v9 = 1;
 		for (;; v8 = (v8 + 1)) {
 			final boolean v11 = (v8 >= 64);
 			if (v11) {
 				break;
 			}
-			final long v15 = (v9 & 16777215);
-			v7[v8] = (int) (v15);
-			final long v20 = (v9 >> 24);
+			final int v15 = (int) (v9 & 16777215);
+			v7[v8] = v15;
+			final int v20 = (int) (v9 >> 24);
 			v9 = v20;
 			continue;
 		}
@@ -32,31 +32,31 @@ public final class examples_big_limbs.java {
 			int[] v30 = null;
 			int[] v35 = v31;
 			int v36 = 0;
-			long v37 = 0;
+			int v37 = 0;
 			for (;; v36 = (v36 + 1)) {
 				final boolean v39 = (v36 >= 64);
 				if (v39) {
 					break;
 				}
 				final boolean v43 = (v36 < 0);
-				long v44 = 0;
+				int v44 = 0;
 				if (v43) {
-					v44 = 0;
+					v44 = (int) (0);
 				} else {
 					final int v48 = v23.length;
 					final boolean v49 = (v36 >= v48);
 					if (v49) {
-						v44 = 0;
+						v44 = (int) (0);
 					} else {
-						final long v55 = v23[v36];
+						final int v55 = v23[v36];
 						v44 = v55;
 					}
 				}
-				final long v56 = (v44 * v24);
+				final long v56 = (((long) v44) * v24);
 				final long v57 = (v56 + v37);
-				final long v59 = (v57 & 16777215);
-				v35[v36] = (int) (v59);
-				final long v64 = (v57 >> 24);
+				final int v59 = (int) (v57 & 16777215);
+				v35[v36] = v59;
+				final int v64 = (int) (v57 >> 24);
 				v37 = v64;
 				continue;
 			}
@@ -74,7 +74,7 @@ public final class examples_big_limbs.java {
 		if (v73) {
 			return 0;
 		}
-		final long v79 = v23[(int) v0];
+		final int v79 = v23[(int) v0];
 		return v79;
 	}
